+++ app/pktgen-capture.h
@@ -60,8 +60,8 @@ typedef struct capture_s {
 	cap_hdr_t                 *tail;/**< Current tailt pointer in the pkt buffer */
 	cap_hdr_t                 *end;	/**< Points to just before the end[-1] of the buffer */
 	size_t used;			/**< Memory used by captured packets */
-	uint8_t lcore;			/**< lcore that captures to this memzone */
-	uint8_t port;			/**< port for this memzone */
+	uint16_t lcore;			/**< lcore that captures to this memzone */
+	uint16_t port;			/**< port for this memzone */
 } capture_t;
 
 /* Capture initialization */
